fix bugs in csv parser related to empty lines in the input.
The solution was to account for the empty lines in handle-char instead
of reinserting the same character and never advancing.  Unfortunately
I needed to add a block in order to make the return value correct,
which makes the relatively simple diff hard to read.

Readline if needed changed from `(= i llen)` to `(>= i llen)` which
prevents heap overflows if our logic is otherwise wrong.

I added this cond to handle-character to prevent it from reinserting the
previous character in the case of empty lines of input.

;; read an empty line, next iteration
`((and
   (member state '(:collecting :collecting-quoted))
   (= i llen 0))
 nil)`

re AccelerationNet/cl-csv#18