Accept *quote* to be nil again.
It's possible to parse CSV files just fine where there's no quote in use,
and it used to be that CL-CSV would be fine with just a setting. Bring back
the old days where it just worked by simply omitting both the quote and the
escape quote when quote is nil.