Don't complete multi-state entries until they're actually done.
When matching a multi-state entry, we need to differentiate between
determining whether the entry matches, and completing that state. A multi-
state entry should only be completed when we've reached the last of its
states. This is accomplished by returning the the completion function or
NIL as the the primary value from CHECK-TABLE-ENTRY, and a secondary value
indicating whether the entry matched.