Do not invoke 'map-fn' on the first line when 'skip-first-p' is true.
Currently, when `:skip-first-p` is true, `:map-fn` is invoked on the first
row and the results are thrown away. In the case where you want to use
the `map-fn` to convert the values to numbers, you will get an error on
the header row. This small patch changes the behavior such that the
`map-fn` is not invoked for the header row if `skip-first-p` is true.

I haven't bene able to run the test suite on my machine, due to an
apparent incompatibility between `asdf` and `lisp-unit2`. I used the
following test case:

```
(defun test-map-fn-with-skip-first-p ()
  (let ((row-count 0))
    (cl-csv:read-csv
     (format nil "skip,header,row~%1,2,3~%")
     :map-fn (lambda (row) (setf row-count (1+ row-count))  row) :skip-first-p t)
    (assert (= row-count 1))))
```